+++ app/api/users.js
@@ -5,11 +5,6 @@ const updateProfileUrl = (userId) => prepareURL(`/api/users/${userId}/update_pro
 const updateAccountUrl = (userId) => prepareURL(`/api/users/${userId}`)
 
 export const getCurrentUser = () => getApi(currentUserURL())
-export const updateProfile = (userId, params) => putApi(updateProfileUrl(userId), {
-  'user': params
-})
-export const updateAccount = (userId, params) => putApi(updateAccountUrl(userId), {
-  'user': params
-})
-
+export const updateProfile = (userId, user) => putApi(updateProfileUrl(userId), { user })
+export const updateAccount = (userId, user) => putApi(updateAccountUrl(userId), { user })
 export const cancelAccount = (userId) => deleteApi(updateAccountUrl(userId))
