+++ app/selectors/index.js
@@ -1,1 +1,2 @@
 export * from './organizations'
+export * from './users'
